UPSTREAM: drm/i915: Evade transcoder's vblank when doing seamless M/N changes
The transcoder M/N values are double buffered on the transcoder's
undelayed vblank. So when doing seamless M/N fastsets we need to
evade also that.

Note that currently the pipe's delayed vblank == transcoder's
undelayed vblank, so this is still a nop change. But in the
future when we may have to delay the pipe's vblank to create
a register programming window ("window2") for the DSB.

Signed-off-by: Ville Syrjälä <ville.syrjala@linux.intel.com>
Link: https://patchwork.freedesktop.org/patch/msgid/20230404175431.23064-2-ville.syrjala@linux.intel.com
Reviewed-by: Mitul Golani <mitulkumar.ajitkumar.golani@intel.com
(cherry picked from commit 435db526a68b6454a882eae7a3768c516d4b540e)

BUG=b:298062999
TEST=DP2.1 displays should work fine

Signed-off-by: Ap, Kamal <kamal.ap@intel.corp-partner.google.com>
Change-Id: I27ce32de00cafb3180af10d6c4d7efb947a390e8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/5054844
Reviewed-by: Sean Paul <sean@poorly.run>
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>
Reviewed-by: Drew Davenport <ddavenport@chromium.org>
Tested-by: Natarajan Balamurugan <balamurugan.natarajan@intel.corp-partner.google.com>