CHROMIUM: add cros specific trace-points.
!!!DO NOT UPSTREAM!!!
!!!DO NOT PORT TO OTHER KERNELS!!!

Add tracepoints to allow the monitoring of the following socket
operations:

sendmsg,recvmsg,listen,accept,release

This is needed because fentry/fexit and LSM BPF hooks are not supported
for ARM64 in kernels older than v6.6.

It has been verified that the v6.6 kernel does not need this patch, both
LSM BPF hooks and fentry/fexit hooks work fine. This is demonstrated by
running the secagentd.NetworkEvents.relaxed_* tests on the strongbad
platform.

BUG=b:329882953
TEST=Ran secagentd.NetworkEvents.relaxed_* on geralt
UPSTREAM-TASK=b:278303728

Change-Id: I2e6e56487566c87b773ddc6ed482fa039603726d
Signed-off-by: Jason Ling <jasonling@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/5441547
Reviewed-by: Pin-yen Lin <treapking@chromium.org>