UPSTREAM: platform/x86/amd: pmc: Report duration of time in hw sleep state
amd_pmc displays a warning when a suspend didn't get to the deepest
state and a dynamic debugging message with the duration if it did.

Rather than logging to dynamic debugging the duration spent in the
deepest state, report this to the standard kernel reporting
infrastructure so that userspace software can query after the
suspend cycle is done.

Reviewed-by: Hans de Goede <hdegoede@redhat.com>
Signed-off-by: Mario Limonciello <mario.limonciello@amd.com>
Signed-off-by: Rafael J. Wysocki <rafael.j.wysocki@intel.com>
(cherry picked from commit 09f5df3fb82fd444296e248f90dd29288e82d3a3)

BUG=b:290556501
TEST=emerge-birman chromeos-kernel-6_1 and dump stb

Change-Id: I7e23f691f5f12c047f6642deccc5900eaced4cbe
Signed-off-by: Harsh Jain <harsh.jain@amd.corp-partner.google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/4677438
Reviewed-by: Sean Paul <sean@poorly.run>
Reviewed-by: Tim Van Patten <timvp@google.com>