CHROMIUM: gpu: mali: Switch to MT8192 upstream binding
The Mali binding on MT8192 was finalized with just a single regulator
and clock. This can be supported on the latest kernel, but we want to
keep using the old schema on older kernels.

This restores CL:4474316 with slight changes where the unnecessary
kernel version based branch are dropped.

BUG=b:222013544
BUG=b:318617295
TEST=Boot to GUI on Asurada with v6.1 kernel
UPSTREAM-TASK=b:260376782

Change-Id: Ie7f932502a83cef9d5cd7b4eb589dca79e89531a
Signed-off-by: Chen-Yu Tsai <wenst@chromium.org>
Signed-off-by: Fei Shao <fshao@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/5173226
Commit-Queue: ChromeOS Auto Retry <chromeos-auto-retry@chromeos-bot.iam.gserviceaccount.com>