UPSTREAM: fbdev/udlfb: Use struct fb_info.screen_buffer
Use info->screen_buffer when reading and writing framebuffers in
system memory. It's the correct pointer for this address space.

The struct fb_info has a union to store the framebuffer memory. This can
either be info->screen_base if the framebuffer is stored in I/O memory,
or info->screen_buffer if the framebuffer is stored in system memory.

As the driver operates on the latter address space, it is wrong to use
.screen_base and .screen_buffer must be used instead. This also gets
rid of casting needed due to not using the correct data type.

Signed-off-by: Thomas Zimmermann <tzimmermann@suse.de>
Reviewed-by: Javier Martinez Canillas <javierm@redhat.com>
Link: https://patchwork.freedesktop.org/patch/msgid/20230428122452.4856-12-tzimmermann@suse.de
(cherry picked from commit 962a3faf32666a31c5a05c2a69023a4abab6389f)
Signed-off-by: Drew Davenport <ddavenport@chromium.org>

BUG=b:307778453
TEST=cq dry run

Backported using forklift.py

Change-Id: I4d5a3605ed0c61f6c53e3033a49df99602c38349