FROMGIT: drm/i915/guc: Print status register when waiting for GuC to load
If the GuC load is taking an excessively long time, the wait loop
currently prints the GT frequency. Extend that to include the GuC
status as well so we can see if the GuC is actually making progress or
not.

Signed-off-by: John Harrison <John.C.Harrison@Intel.com>
Reviewed-by: Daniele Ceraolo Spurio <daniele.ceraolospurio@intel.com>
Link: https://patchwork.freedesktop.org/patch/msgid/20230502234007.1762014-3-John.C.Harrison@Intel.com
(cherry picked from commit c354feb5895fe80fd3f896443015cf53d5cf98e8
 https://anongit.freedesktop.org/git/drm-intel.git drm-intel-gt-next)

BUG=b:265114629
TEST=i915 driver loaded successfully

Change-Id: Ib28d673b4f6aa2472e186cc36f8dfbc04cdcdff9
Signed-off-by: Ap, Kamal <kamal.ap@intel.corp-partner.google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/4511857
Reviewed-by: Drew Davenport <ddavenport@chromium.org>
Tested-by: Waheem Md <waheem.md@intel.com>
Reviewed-by: Sean Paul <sean@poorly.run>