UPSTREAM: sched/topology: Propagate SMT flags when removing degenerate domain
When a degenerate cluster domain for core with SMT CPUs is removed,
the SD_SHARE_CPUCAPACITY flag in the local child sched group was not
propagated to the new parent.  We need this flag to properly determine
whether the local sched group is SMT.  Set the flag in the local
child sched group of the new parent sched domain.

Signed-off-by: Tim Chen <tim.c.chen@linux.intel.com>
Signed-off-by: Peter Zijlstra (Intel) <peterz@infradead.org>
Reviewed-by: Ricardo Neri <ricardo.neri-calderon@linux.intel.com>
Link: https://lkml.kernel.org/r/73cf0959eafa53c02e7ef6bf805d751d9190e55d.1683156492.git.tim.c.chen@linux.intel.com
(cherry picked from commit bf2dc42d6beb890c995b8b09f881ef1b37259107)

BUG=b:242115011
TEST=Build and Boot Sanity

Change-Id: Ic52dcd215fdf268e7465d842859a73c5a5e02c8c
Signed-off-by: Sachin Kumar <sachin.kumar@intel.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/4866018
Tested-by: Basavaprasad Kanshetty <basavaprasad.kanshetty@intel.corp-partner.google.com>
Reviewed-by: Paz Zcharya <pazz@google.com>
Reviewed-by: Sean Paul <sean@poorly.run>
Tested-by: Adeel Arshad <adeel.arshad@intel.corp-partner.google.com>